DECLARAMOS FUNCIONES EN MainActivity.kt

+++ app/src/main/java/com/example/calculadora/Calculadora.kt
@@ -1,15 +1,13 @@
 package com.example.calculadora
 
-class Calculadora
-{
-    fun suma(a: Int, b: Int): Int = a + b
-    fun resta(a: Int, b: Int): Int = a - b
-    fun multiplicacion(a: Int, b: Int): Int = a * b
-    fun division(a: Int, b: Int): Double
-    {
-        if (b == 0) {
+class Calculadora {
+    fun suma(a: Double, b: Double): Double = a + b
+    fun resta(a: Double, b: Double): Double = a - b
+    fun multiplicacion(a: Double, b: Double): Double = a * b
+    fun division(a: Double, b: Double): Double {
+        if (b == 0.0) {
             throw IllegalArgumentException("No se puede dividir entre 0")
         }
-        return a.toDouble() / b
+        return a / b
     }
 }
